Metier, irepo, metier Phase - maryssa

+++ Projet_PAF/src/main/java/com/inti/formation/IMetier/IPhaseMetier.java
@@ -2,6 +2,6 @@ package com.inti.formation.IMetier;
 
 import com.inti.formation.Model.Phase;
 
-public interface IPhaseMetier extends IGenericMetier<Phase, Integer> {
+public interface IPhaseMetier extends IGenericMetier<Phase, Long> {
 
 }
